+++ client/src/pages/create-checker-page/CustomiseColors.tsx
@@ -2,7 +2,8 @@ import styles from "./style.module.css";
 import Button from "../../components/Button";
 import BackButton from "../../components/BackButton";
 import { useContext, Dispatch, SetStateAction } from "react";
-import { PageContextProvider, Page } from "./CreateCheckerPage";
+import { PageContextProvider } from "./CreateCheckerPage";
+import Page from "../../types/Page";
 import { ColorPicker } from "./CustomiseColors Components/ColorPicker";
 import { ChromePicker } from "react-color";
 import ClubCheckerPage from "../club-checker-page/ClubCheckerPage";
