+++ client/src/pages/create-checker-page/CustomiseFont.tsx
@@ -3,7 +3,8 @@ import styles from "./style.module.css";
 import Button from "../../components/Button";
 import BackButton from "../../components/BackButton";
 import { Dispatch, SetStateAction, useContext } from "react";
-import { PageContextProvider, Page } from "./CreateCheckerPage";
+import { PageContextProvider } from "./CreateCheckerPage";
+import Page from "../../types/Page";
 import FontPicker from "react-fontpicker-ts";
 import "react-fontpicker-ts/dist/index.css";
 import { ArrowDown2 } from "iconsax-react";
