+++ client/src/pages/create-checker-page/CustomiseLogo.tsx
@@ -2,7 +2,8 @@ import styles from "./style.module.css";
 import BackButton from "../../components/BackButton";
 import Button from "../../components/Button";
 import { useContext, Dispatch, SetStateAction } from "react";
-import { PageContextProvider, Page } from "./CreateCheckerPage";
+import { PageContextProvider } from "./CreateCheckerPage";
+import Page from "../../types/Page";
 import UploadButton from "../../components/UploadButton";
 import ClubCheckerPage from "../club-checker-page/ClubCheckerPage";
