+++ client/src/pages/create-checker-page/GoogleSheetForm.tsx
@@ -8,7 +8,8 @@ import {
   useEffect,
   useState,
 } from "react";
-import { PageContextProvider, Page } from "./CreateCheckerPage";
+import { PageContextProvider } from "./CreateCheckerPage";
+import Page from "../../types/Page";
 import Textfield from "../../components/Textfield";
 import Button from "../../components/Button";
 import LinkIcon from "../../assets/LinkIcon.svg";
